add demo and gifs
