+++ .babelrc
@@ -9,4 +9,3 @@
     '@babel/preset-react'
   ]
 }
-
