+++ package.json
@@ -33,7 +33,6 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "regenerator-runtime": "^0.13.9",
-    "sqlite3": "^5.0.2",
     "tailwindcss": "^2.2.4",
     "util": "^0.12.4"
   },
