--- src/infiniteScroll.js
@@ -1,28 +1,0 @@
-// import React, {useState, useEffect} from 'react';
-
-// const useInfiniteScroll = (callback) => {
-//   const [isFetching, setIsFetching] = useState(false);
-
-//   useEffect(() => {
-//     window.addEventListener('scroll', handleScroll);
-//     return () => window.removeEventListener('scroll', handleScroll);
-//   }, []);
-
-//   useEffect(() => {
-//     if (!isFetching) return;
-//     callback()
-//   }, [isFetching]);
-
-//   const handleScroll = () => {
-//     if (window.innerHeight + document.documentElement.scrollTop !==
-//        document.documentElement.offsetHeight || isFetching) return;
-//     setIsFetching(true);
-//   }
-
-//  return [isFetching, setIsFetching];
-
-// };
-
-// export default useInfiniteScroll;
-
-
