+++ src/styles.css
@@ -22,7 +22,6 @@
   background-color: (248, 240, 227);
 }
 
-
 .cardHeading {
   font-family: sans-serif;
   font-size: 20px;
