+++ webpack.config.js
@@ -14,20 +14,20 @@ const config = {
         use: 'babel-loader',
         exclude: /node_modules/
       },
-      // {
-      //   test: /\.css$/,
-      //   use: [
-      //     'style-loader',
-      //     {
-      //       loader: 'css-loader',
-      //       options: {
-      //         importLoaders: 1
-      //       }
-      //     },
-      //     'postcss-loader'
-      //   ],
-      //   exclude: /\.module\.css$/
-      // },
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          {
+            loader: 'css-loader',
+            options: {
+              importLoaders: 1
+            }
+          },
+          'postcss-loader'
+        ],
+        exclude: /\.module\.css$/
+      },
       {
         test: /\.css$/,
         use: [
